tests/wpt/tests/content-security-policy/style-src-attr-elem/style-src-elem-allowed-src-blocked-link.html

<!DOCTYPE html>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="style-src 'none'; style-src-elem 'self'">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      var t = async_test("Inline style should be applied");
      window.addEventListener('securitypolicyviolation', t.unreached_func("Should not have fired a spv event"));
    </script>
    <link rel="stylesheet" href="/content-security-policy/style-src/resources/allowed.css">
</head>
<body>
    <script>
      t.step(function() {
        assert_equals(document.styleSheets.length, 1);
        t.done();
      });
    </script>
</body>

</html>
